refactor: change global presentation folder name to global controller
