+++ src/main/java/com/opensource/hongroaddong/global/controller/HealthController.java
@@ -1,4 +1,4 @@
-package com.opensource.hongroaddong.global.presentation;
+package com.opensource.hongroaddong.global.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
